backend/requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
Django==4.2.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2024.1
sqlparse==0.4.4
